Added Tour Review Page

+++ Model/TourReview.cs
@@ -81,7 +81,7 @@ namespace BookingApp.Model
             KnowledgeGrade = int.Parse(values[1]);
             LanguageGrade = int.Parse(values[2]);
             InterestGrade = int.Parse(values[3]);
-            AvrageGrade = int.Parse(values[4]);
+            AvrageGrade = double.Parse(values[4]);
             Posted = DateTime.ParseExact(values[5], "dd-MM-yyyy HH:mm", CultureInfo.InvariantCulture);
             IsValid = bool.Parse(values[6]);
             TouristId = int.Parse(values[7]);
